refactor(rum-explorer): move custom elements into separate folder
