+++ tools/rum/elements/incognito-checkbox.js
