+++ tools/rum/flow.html
@@ -12,9 +12,9 @@
   <link rel="stylesheet" href="/styles/styles.css" />
   <script type="module" defer="false">
     import SankeyChart from './charts/sankey.js';
-    import IncognitoCheckbox from './incognito-checkbox.js';
-    import FacetSidebar from './facetsidebar.js';
-    import ListFacet from './list-facet.js';
+    import IncognitoCheckbox from './elements/incognito-checkbox.js';
+    import FacetSidebar from './elements/facetsidebar.js';
+    import ListFacet from './elements/list-facet.js';
     window.slicer = {
       Chart: SankeyChart,
     };
