+++ tools/rum/list.html
@@ -11,9 +11,9 @@
   <link rel="stylesheet" href="/styles/styles.css" />
   <script type="module" defer="false">
   import BarChart from './charts/barchart.js';
-  import IncognitoCheckbox from './incognito-checkbox.js';
-  import FacetSidebar from './facetsidebar.js';
-  import ListFacet from './list-facet.js';
+  import IncognitoCheckbox from './elements/incognito-checkbox.js';
+  import FacetSidebar from './elements/facetsidebar.js';
+  import ListFacet from './elements/list-facet.js';
   window.slicer = {
     Chart: BarChart,
   };
